\name{cnSentimentR-package}
\alias{cnSentimentR-package}
\alias{cnSentimentR}
\docType{package}
\title{
\packageTitle{cnSentimentR}
}
\description{
\packageDescription{cnSentimentR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cnSentimentR}
\packageIndices{cnSentimentR}
}
\author{
\packageAuthor{cnSentimentR}

Maintainer: \packageMaintainer{cnSentimentR}
}
\references{
	URL: https://github.com/leeshuheng/cnSentimentR
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
	cnsr.predict,cnsr.train,cnsr.prepare
}
